+++ source/core/src/main/com/csse3200/game/areas/terrain/TerrainComponent.java
@@ -243,340 +243,4 @@ public class TerrainComponent extends RenderComponent {
     ISOMETRIC,
     HEXAGONAL
   }
-
-  /**
-   * TerrainResource class to store all possible tiles and their edge tiles.
-   */
-  public class TerrainResource {
-    private List<Tile> forestTiles;
-    private List<Tile> waterTiles;
-    private List<Tile> airTiles;
-    private List<Tile> fogTiles;
-
-    // total number of each tile
-    public static int FOREST_SIZE = 0;
-    public static int WATER_SIZE = 0;
-    public static int AIR_SIZE = 0;
-    public static int FOG_SIZE = 0;
-    private boolean unlockedWater;
-
-    public TerrainResource(MapType mapType) {
-      ResourceService resourceService = ServiceLocator.getResourceService();
-      forestTiles = new ArrayList<>();
-      waterTiles = new ArrayList<>();
-      airTiles = new ArrayList<>();
-      fogTiles = new ArrayList<>();
-      this.unlockedWater = false;
-      switch(mapType) {
-        case FOREST:
-          // load forest tiles
-          for (TileConfig tile : ForestTileConfig.forestMapTiles) {
-            // edge: TOP, RIGHT, BOTTOM, LEFT
-            // A: sand, B: grass, C: water
-            // =======================
-            forestTiles.add(
-                new Tile(tile.id, 
-                        new TextureRegion(resourceService.getAsset(tile.fp, Texture.class)),
-                        tile.edges, tile.centre));
-          }
-          FOREST_SIZE = forestTiles.size();
-
-          // load water tiles
-          OceanMapTiles oceanTileConfig;
-          FogMapTiles fogTileConfig;
-          oceanTileConfig = FileLoader.readClass(OceanMapTiles.class, "configs/OceanGameAreaConfigs/waterTiles.json");
-          fogTileConfig = FileLoader.readClass(FogMapTiles.class, "configs/FogGameAreaConfigs/fogTiles.json");
-
-          for (OceanTileConfig tile : oceanTileConfig.waterMapTiles) {
-            waterTiles.add(new Tile(tile.id,
-                    new TextureRegion(resourceService.getAsset(tile.fp, Texture.class)),
-                    tile.edges,
-                    tile.centre));
-          }
-          WATER_SIZE = waterTiles.size();
-
-          // load fog tiles
-          for (FogTileConfig tile : fogTileConfig.fogTiles) {
-            fogTiles.add(new Tile(tile.id,
-                    new TextureRegion(resourceService.getAsset(tile.fp, Texture.class)),
-                    tile.edges,
-                    tile.centre));
-          }
-          FOG_SIZE = fogTiles.size();
-
-          break;
-        case COMBAT:
-          break;
-        case MAZE_MINIGAME:
-          break;
-        default:
-          throw new IllegalArgumentException("Map type not supported: " + mapType);
-      }
-
-      this.setPossibleTiles();
-    }
-
-    public List<Tile> getMapTiles(MapType mapType) {
-      switch(mapType) {
-        case FOREST:
-          return forestTiles;
-        case WATER:
-          return waterTiles;
-        case AIR:
-          return airTiles;
-        case FOG:
-          return fogTiles;
-        case COMBAT:
-          return null;
-        default:
-          throw new IllegalArgumentException("No such map type:" + mapType);
-      }
-    }
-
-    /**
-     * Set all possible tiles for each tile for all directions.
-     */
-    public void setPossibleTiles() {
-      for (int i = 0; i < this.forestTiles.size(); i++) {
-        setPossibleUp(this.forestTiles.get(i), this.forestTiles);
-        setPossibleRight(this.forestTiles.get(i), this.forestTiles);
-        setPossibleDown(this.forestTiles.get(i), this.forestTiles);
-        setPossibleLeft(this.forestTiles.get(i), this.forestTiles);
-      }
-
-      for (int i = 0; i < this.waterTiles.size(); i++) {
-        setPossibleUp(this.waterTiles.get(i), this.waterTiles);
-        setPossibleRight(this.waterTiles.get(i), this.waterTiles);
-        setPossibleDown(this.waterTiles.get(i), this.waterTiles);
-        setPossibleLeft(this.waterTiles.get(i), this.waterTiles);
-      }
-
-      for (int i = 0; i < this.fogTiles.size(); i++) {
-        setPossibleUp(this.fogTiles.get(i), this.fogTiles);
-        setPossibleRight(this.fogTiles.get(i), this.fogTiles);
-        setPossibleDown(this.fogTiles.get(i), this.fogTiles);
-        setPossibleLeft(this.fogTiles.get(i), this.fogTiles);
-      }
-    }
-
-    /**
-     * Set possible tiles for up direction.
-     * 
-     * @param tile The tile to set possible tiles
-     */
-    public void setPossibleUp(Tile tile, List<Tile> areaTiles) {
-      BitSet up = new BitSet(areaTiles.size());
-      for (int i = 0; i < areaTiles.size(); i++) {
-        if (areaTiles.get(i).getEdgeTiles().get(2).equals(tile.getEdgeTiles().get(0))) {
-          up.set(i, true);
-        }
-      }
-      tile.setPossibleUp(up);
-    }
-
-    /**
-     * Set possible tiles for right direction.
-     * 
-     * @param tile The tile to set possible tiles
-     */
-    public void setPossibleRight(Tile tile, List<Tile> areaTiles) {
-      BitSet right = new BitSet(areaTiles.size());
-      for (int i = 0; i < areaTiles.size(); i++) {
-        if (areaTiles.get(i).getEdgeTiles().get(3).equals(tile.getEdgeTiles().get(1))) {
-          right.set(i, true);
-        }
-      }
-      tile.setPossibleRight(right);
-    }
-
-    /**
-     * Set possible tiles for down direction.
-     * 
-     * @param tile The tile to set possible tiles
-     */
-    public void setPossibleDown(Tile tile, List<Tile> areaTiles) {
-      BitSet down = new BitSet(areaTiles.size());
-      for (int i = 0; i < areaTiles.size(); i++) {
-        if (areaTiles.get(i).getEdgeTiles().get(0).equals(tile.getEdgeTiles().get(2))) {
-          down.set(i, true);
-        }
-      }
-      tile.setPossibleDown(down);
-    }
-
-    /**
-     * Set possible tiles for left direction.
-     * 
-     * @param tile The tile to set possible tiles
-     */
-    public void setPossibleLeft(Tile tile, List<Tile> areaTiles) {
-      BitSet left = new BitSet(areaTiles.size());
-      for (int i = 0; i < areaTiles.size(); i++) {
-        if (areaTiles.get(i).getEdgeTiles().get(1).equals(tile.getEdgeTiles().get(3))) {
-          left.set(i, true);
-        }
-      }
-      tile.setPossibleLeft(left);
-    }
-
-    /**
-     * Get a tile by index and map type.
-     * 
-     * @param index The index of the tile
-     * @param mapType The map type of the tile
-     * @return The tile with the given index and map type
-     */
-    public Tile getMapTilebyIndex(int index, MapType mapType) {
-      switch(mapType) {
-        case FOREST:
-          return forestTiles.get(index);
-        case FOG:
-          return fogTiles.get(index);
-        case WATER:
-          return waterTiles.get(index);
-        case AIR:
-          return airTiles.get(index);
-        case COMBAT:
-          return null;
-        default:
-          throw new IllegalArgumentException("No such map type:" + mapType);
-      }
-    }
-  }
-
-  /**
-   * Tile class to store tile data.
-   * A tile has a name, texture, and edge tiles.
-   */
-  public class Tile {
-    // data for wave function collapse
-    private TextureRegion texture;
-    private List<String> edgeTiles;
-    private String name;
-    private String centre;
-
-    // all possible tiles
-    private BitSet up = new BitSet();
-    private BitSet right = new BitSet();
-    private BitSet down = new BitSet();
-    private BitSet left = new BitSet();
-
-    public boolean collapsed = false;
-
-    public Tile(String name, TextureRegion texture, List<String> edgeTiles, String centre) {
-      this.name = name;
-      this.texture = texture;
-      this.edgeTiles = edgeTiles;
-      this.centre = centre;
-    }
-
-    /**
-     * Get the name of the tile.
-     * 
-     * @return The name of the tile
-     */
-    public String getName() {
-      return name;
-    }
-
-    /**
-     * Get the texture of the tile.
-     * 
-     * @return The texture of the tile
-     */
-    public TextureRegion getTexture() {
-      return texture;
-    }
-
-    /**
-     * Get the edge tiles of the tile.
-     * 
-     * @return The edge tiles of the tile
-     */
-    public List<String> getEdgeTiles() {
-      return edgeTiles;
-    }
-
-    /**
-     * Get the centre tile of the tile.
-     * 
-     * @return The centre tile of the tile
-     */
-    public String getCentre() {
-      return centre;
-    }
-
-    /**
-     * Get the up edge tiles of the tile.
-     * 
-     * @return The up edge tiles of the tile
-     */
-    public BitSet getUp() {
-      return up;
-    }
-
-    /**
-     * Get the right edge tiles of the tile.
-     * 
-     * @return The right edge tiles of the tile
-     */
-    public BitSet getRight() {
-      return right;
-    }
-
-    /**
-     * Get the down edge tiles of the tile.
-     * 
-     * @return The down edge tiles of the tile
-     */
-    public BitSet getDown() {
-      return down;
-    }
-
-    /**
-     * Get the left edge tiles of the tile.
-     * 
-     * @return The left edge tiles of the tile
-     */
-    public BitSet getLeft() {
-      return left;
-    }
-
-    /**
-     * Set the possible up edge tiles of the tile.
-     * 
-     * @param up All up edge tiles of the tile
-     */
-    public void setPossibleUp(BitSet up) {
-      this.up = up;
-    }
-
-    /**
-     * Set the possible right edge tiles of the tile.
-     * 
-     * @param right All right edge tiles of the tile
-     */
-    public void setPossibleRight(BitSet right) {
-      this.right = right;
-    }
-
-    /**
-     * Set the possible down edge tiles of the tile.
-     * 
-     * @param down All down edge tiles of the tile
-     */
-    public void setPossibleDown(BitSet down) {
-      this.down = down;
-    }
-
-    /**
-     * Set the possible left edge tiles of the tile.
-     * 
-     * @param left All left edge tiles of the tile
-     */
-    public void setPossibleLeft(BitSet left) {
-      this.left = left;
-    }
-  }
-
 }
